requirements.txt
flask
praw
python-dotenv
openai